Structured Views and added SubViewModel, UpdateSVM and IoSVM
