+++ PixiEditor/Views/Dialogs/ConfirmationPopup.xaml
