+++ PixiEditor/Views/Dialogs/ConfirmationPopup.xaml.cs
