+++ PixiEditor/Views/Dialogs/ImportFilePopup.xaml
