+++ PixiEditor/Views/Dialogs/ImportFilePopup.xaml.cs
