+++ PixiEditor/Views/Dialogs/NewFilePopup.xaml
