+++ PixiEditor/Views/Dialogs/NewFilePopup.xaml.cs
