+++ PixiEditor/Views/Dialogs/PopupTemplate.xaml
