+++ PixiEditor/Views/Dialogs/PopupTemplate.xaml.cs
