+++ PixiEditor/Views/Dialogs/ResizeCanvasPopup.xaml
