+++ PixiEditor/Views/Dialogs/ResizeCanvasPopup.xaml.cs
