+++ PixiEditor/Views/Dialogs/ResizeDocumentPopup.xaml
