+++ PixiEditor/Views/Dialogs/ResizeDocumentPopup.xaml.cs
