+++ PixiEditor/Views/UserControls/AnchorPointPicker.xaml
