+++ PixiEditor/Views/UserControls/AnchorPointPicker.xaml.cs
