+++ PixiEditor/Views/UserControls/LayerItem.xaml
