+++ PixiEditor/Views/UserControls/LayerItem.xaml.cs
