+++ PixiEditor/Views/UserControls/MainDrawingPanel.xaml
