+++ PixiEditor/Views/UserControls/MainDrawingPanel.xaml.cs
