+++ PixiEditor/Views/UserControls/MenuButton.xaml
