+++ PixiEditor/Views/UserControls/NumberInput.xaml
