+++ PixiEditor/Views/UserControls/NumberInput.xaml.cs
