+++ PixiEditor/Views/UserControls/Rotatebox.xaml
