+++ PixiEditor/Views/UserControls/Rotatebox.xaml.cs
