+++ PixiEditor/Views/UserControls/SizeInput.xaml
