+++ PixiEditor/Views/UserControls/SizeInput.xaml.cs
