+++ PixiEditor/Views/UserControls/SizePicker.xaml
