+++ PixiEditor/Views/UserControls/SizePicker.xaml.cs
